requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
sqlite-vec>=0.1.1
playwright>=1.41.0
